refactor(jwt): replace jsonwebtoken with jwt-compact for cross-platform support
- Replace jsonwebtoken with jwt-compact library
- Unify JWT implementation for native and WASM platforms
- Remove mutually exclusive crypto backend features
- Reduce JWT code complexity by ~50% (300→150 lines)
- Fix example code compilation errors
- Add comprehensive JWT test suite (14 tests)
- Update ISSUES.md and add 008_jwt_revise.md plan
- Fix clippy warnings in JWT module

Breaking Changes:
- Internal JWT implementation replaced
- No public API changes for users

Closes: JSONwebtoken compilation error with duplicate definitions
Related: WASM support implementation

Files modified:
- Cargo.toml: Update dependencies and features
- src/internal/jwt.rs: Complete rewrite with jwt-compact
- src/auth.rs: Updated to use new JWT module
- src/error.rs: Update JWT error type
- src/lib.rs: Remove unused code
- examples/github_app_authentication.rs: Fix build issue
- examples/github_app_authentication_manual.rs: Simplify usage
- tests/jwt_test.rs: Add comprehensive tests
- ISSUES.md: Update status and document completion
- plans/008_jwt_revise.md: Add completion plan

+++ src/auth.rs
@@ -67,7 +67,7 @@ pub fn create_jwt(github_app_id: AppId, key: &jwt::EncodingKey) -> Result<String
                 backtrace: snafu::Backtrace::capture(),
             }
         })?
-        .as_secs();
+        .as_secs() as i64;
 
     // Github only allows JWTs that expire in the next 10 minutes.
     // The token is issued 60 seconds in the past and expires in 9 minutes,
